Test failure for endpoint and flow models.

+++ src/main/java/netlab/submission/request/Request.java
@@ -31,7 +31,7 @@ public class Request {
 
     private Map<SourceDestPair, Map<String, Path>> chosenPaths;
 
-    private Long runningTimeMillis;
+    private double runningTimeMillis;
 
     private Boolean isFeasible;
 }
